+++ noodles-util/src/variant.rs
@@ -3,7 +3,8 @@
 #[cfg(feature = "async")]
 pub mod r#async;
 
+mod index;
 pub mod io;
 mod record;
 
-pub use self::record::Record;
+pub use self::{index::Index, record::Record};