+++ noodles-util/src/variant/index.rs
@@ -1,0 +1,10 @@
+use noodles_csi as csi;
+use noodles_vcf as vcf;
+
+/// A variant index.
+pub enum Index {
+    /// A VCF index.
+    Vcf(vcf::Index),
+    /// A BCF index.
+    Bcf(csi::Index),
+}